Switch to v4 and use consistent version naming
