+++ crates/sovereign-sdk/rollup-interface/src/state_machine/zk/batch_proof/input/v4.rs
@@ -1,10 +1,51 @@
-use std::collections::VecDeque;
+use borsh::{BorshDeserialize, BorshSerialize};
 
-use super::v3::{BatchProofCircuitInputV3Part1, BatchProofCircuitInputV3Part2, PrevHashProof};
-use crate::block::L2Block;
+use crate::block::{L2Block, L2Header};
 use crate::da::SequencerCommitment;
 use crate::witness::Witness;
 use crate::zk::StorageRootHash;
+use std::collections::VecDeque;
+
+#[derive(BorshDeserialize, BorshSerialize)]
+// Prevent serde from generating spurious trait bounds. The correct serde bounds are already enforced by the
+// StateTransitionFunction, DA, and Zkvm traits.
+/// First part of the v4 elf input
+pub struct BatchProofCircuitInputV4Part1 {
+    /// The state root before the state transition
+    pub initial_state_root: StorageRootHash,
+    /// The sequencer commitment before the first sequencer commitment in the sequencer_commitments vector
+    /// If it is none than this is the first batch proof
+    /// Else the index of the sequencer commitment should be `sequencer_commitments[0].index - 1``
+    pub previous_sequencer_commitment: Option<SequencerCommitment>,
+    /// A proof for the previous sequencer commitment's last header
+    /// None if the previous sequencer commitment is None
+    pub prev_hash_proof: Option<PrevHashProof>,
+    /// Sequencer commitments being proven
+    /// Since `SequencerCommitment` does not have the sequencer's signature,
+    /// the light client prover will be doing the signature verification
+    /// when it is extracting the commitments from L1
+    pub sequencer_commitments: Vec<SequencerCommitment>,
+    /// Short header proofs for verifying system transactions
+    pub short_header_proofs: VecDeque<Vec<u8>>,
+    /// L2 heights in which the guest should prune the log caches to avoid OOM.
+    pub cache_prune_l2_heights: Vec<u64>,
+    /// The witness needed to access the last L1 hash on the bitcoin light client contract
+    pub last_l1_hash_witness: Witness,
+}
+
+#[derive(BorshDeserialize, BorshSerialize)]
+/// A merkle proof for the last header in the previous sequencer commitment
+/// This is used to verify the first `prev_hash` in the batch proof circuit
+/// The `prev_hash` is the hash of the last header in the previous sequencer commitment
+pub struct PrevHashProof {
+    /// Rightmost header in the L2 block hash merkle tree
+    pub last_header: L2Header,
+    /// Merkle proof for the last header in the previous sequencer commitment
+    pub merkle_proof_bytes: Vec<u8>,
+    /// Give the start of the previous sequencer commitment as a hint
+    /// so index can be calculated
+    pub prev_sequencer_commitment_start: u64,
+}
 
 /// Pre-computed ecrecover pubkey witnesses, grouped by sequencer commitment.
 ///
@@ -12,22 +53,35 @@ use crate::zk::StorageRootHash;
 /// signatures of the L2 blocks in that commitment, in circuit consumption order.
 pub type EcrecoverPubkeyWitnesses = VecDeque<Vec<[u8; 65]>>;
 
+#[derive(BorshDeserialize, BorshSerialize)]
+/// Second part of the v4 elf input
+/// This is going to be read per-need basis to not go out of memory
+/// in the zkvm
+pub struct BatchProofCircuitInputV4Part2(pub EcrecoverPubkeyWitnesses);
+
+type InputV4Part3<Witness> = VecDeque<Vec<(u64, L2Block, Witness, Witness)>>;
+
+#[derive(BorshDeserialize, BorshSerialize)]
+/// Third part of the v4 elf input
+/// This is going to be read per-need basis to not go out of memory
+/// in the zkvm
+pub struct BatchProofCircuitInputV4Part3(pub InputV4Part3<Witness>);
+
 /// Legacy batch-proof input layout.
-pub type LegacyBatchProofInput = (BatchProofCircuitInputV3Part1, BatchProofCircuitInputV3Part2);
+pub type LegacyBatchProofInput = (BatchProofCircuitInputV4Part1, BatchProofCircuitInputV4Part3);
 
 /// Batch-proof input V4 layout.
 pub type BatchProofCircuitInputV4Parts = (
-    BatchProofCircuitInputV3Part1,
-    EcrecoverPubkeyWitnesses,
-    BatchProofCircuitInputV3Part2,
+    BatchProofCircuitInputV4Part1,
+    BatchProofCircuitInputV4Part2,
+    BatchProofCircuitInputV4Part3,
 );
 
-/// Host-side batch-proof input data for the V4 layout.
-///
-/// Reuses the V3 `Part1`/`Part2` parts on the wire and adds, for forks that
-/// activate it, one pre-computed ecrecover pubkey-witness stream item between
-/// them. The host serializes this either as the legacy V3 input or as the
-/// pubkey-witness input, depending on the active fork.
+#[derive(BorshDeserialize, BorshSerialize)]
+// Prevent serde from generating spurious trait bounds. The correct serde bounds are already enforced by the
+// StateTransitionFunction, DA, and Zkvm traits.
+/// Data required to verify a state transition.
+/// This is more like a glue type to create V1/V2/V3 batch proof circuit inputs later in the program
 pub struct BatchProofCircuitInputV4 {
     /// The state root before the state transition
     pub initial_state_root: StorageRootHash,
@@ -60,12 +114,12 @@ impl BatchProofCircuitInputV4 {
     fn into_shared_parts(
         self,
     ) -> (
-        BatchProofCircuitInputV3Part1,
-        EcrecoverPubkeyWitnesses,
-        BatchProofCircuitInputV3Part2,
+        BatchProofCircuitInputV4Part1,
+        BatchProofCircuitInputV4Part2,
+        BatchProofCircuitInputV4Part3,
     ) {
         assert_eq!(self.l2_blocks.len(), self.state_transition_witnesses.len());
-        let mut part2 = VecDeque::with_capacity(self.l2_blocks.len());
+        let mut part3 = VecDeque::with_capacity(self.l2_blocks.len());
 
         for (l2_blocks, witnesses) in self
             .l2_blocks
@@ -82,10 +136,10 @@ impl BatchProofCircuitInputV4 {
                 })
                 .collect();
 
-            part2.push_back(v);
+            part3.push_back(v);
         }
 
-        let part1 = BatchProofCircuitInputV3Part1 {
+        let part1 = BatchProofCircuitInputV4Part1 {
             initial_state_root: self.initial_state_root,
             previous_sequencer_commitment: self.previous_sequencer_commitment,
             prev_hash_proof: self.prev_hash_proof,
@@ -97,16 +151,16 @@ impl BatchProofCircuitInputV4 {
 
         (
             part1,
-            self.ecrecover_pubkey_witnesses,
-            BatchProofCircuitInputV3Part2(part2),
+            BatchProofCircuitInputV4Part2(self.ecrecover_pubkey_witnesses),
+            BatchProofCircuitInputV4Part3(part3),
         )
     }
 
     /// Build the legacy input layout. Pubkey witnesses are intentionally
     /// dropped because older circuits recover pubkeys in-VM.
     pub fn into_legacy_parts(self) -> LegacyBatchProofInput {
-        let (part1, _ecrecover_pubkey_witnesses, part2) = self.into_shared_parts();
-        (part1, part2)
+        let (part1, _part2, part3) = self.into_shared_parts();
+        (part1, part3)
     }
 
     /// Build the V4 input layout.
